Cleaning rooms controller by creating a helper
